[수정] 분리한 커스텀훅 등을 각각의 컴포넌트로 이동
